# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Pistas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/api/get-comments/{court_id}": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Obtener todos los comentarios de una pista",
                "operationId": "5331ce82bdb7dcb375fafc86b5a8d210",
                "parameters": [
                    {
                        "name": "court_id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Los comentarios se han obtenido correctamente\n    *           No hay comentarios de esta pista"
                    },
                    "404": {
                        "description": "\n    *           La pista no existe"
                    }
                }
            }
        },
        "/api/add-comment/{court_id}": {
            "post": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Añadir un comentario",
                "operationId": "52f6b721971d73e3d7b5a5e0aef1dad3",
                "parameters": [
                    {
                        "name": "court_id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "texto",
                        "in": "query",
                        "description": "Texto del comentario a añadir",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "like",
                        "in": "query",
                        "description": "Este parámetro representa si el comentario es positivo o negativo",
                        "required": true,
                        "schema": {
                            "type": "bool"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha publicado el comentario correctamente"
                    },
                    "400": {
                        "description": "\n    *           Error de validación"
                    },
                    "404": {
                        "description": "\n    *           No existe la pista donde se quiere publicar el comentario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-comment/{id}": {
            "delete": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Eliminar un comentario",
                "operationId": "4f992bb7f74f1c4a6be77177c01c545b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha borrado el comentario correctamente"
                    },
                    "403": {
                        "description": "\n    *           Debes ser Administrador, Moderador o propietario de este comentario para poder borrarlo"
                    },
                    "404": {
                        "description": "\n    *           Este comentario no existe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}